Showing the note in the console
