+++ alba-notes/src/firebase/create-collection.js
@@ -1,25 +1,29 @@
 
 import db from "./firebase-init";
 import { collection, addDoc } from "firebase/firestore";
-import {  getDocs } from "firebase/firestore";
+//import {  getDocs } from "firebase/firestore";
+import { useState } from "react";
+
+
 
 export const AddingNotes = async(e) =>{
+    
 try {
-  const docRef = await addDoc(collection(db, "users"), {
-    first: "Ada",
-    last: "Lovelace",
-    born: 1815
+  const docRef = await addDoc(collection(db, "notes"), {
+    title: "Alo",
+    note: "Polesia"
   });
+  console.log('Working') 
   console.log("Document written with ID: ", docRef.id);
 } catch (e) {
   console.error("Error adding document: ", e);
 }
 }
 
-export const ViewNotes = async(e) =>{
+// export const ViewNotes = async(e) =>{
 
-const querySnapshot = await getDocs(collection(db, "users"));
-querySnapshot.forEach((doc) => {
-  console.log(`${doc.id} => ${doc.data()}`);
-});
-}
+// const querySnapshot = await getDocs(collection(db, "notes"));
+// querySnapshot.forEach((doc) => {
+//   console.log(`${doc.id} => ${doc.data()}`);
+// });
+// }
